// app/src/generated/UsageOdds.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * An initializer's `usage_odds`.
 */
export type UsageOdds = { "kind": "number", value: number, } | { "kind": "unknown", written: string, } | { "kind": "script", text: string, base: number | null, };